+++ fabric/src/main/java/dev/thenu/ReddensStoneLanternReconstructed/init/BlockFile.java
@@ -73,8 +73,8 @@ public class BlockFile {
     public static Block BROAD_STONE_LANTERN_BLOCK_LIGHT;
     public static Block MID_STONE_LANTERN_BLOCK_DARK;
     public static Block MID_STONE_LANTERN_BLOCK_LIGHT;
-    public static Block THIN_STONE_LANTERN_BLOCK_DARK;
-    public static Block THIN_STONE_LANTERN_BLOCK_LIGHT;
+    /*public static Block THIN_STONE_LANTERN_BLOCK_DARK;
+    public static Block THIN_STONE_LANTERN_BLOCK_LIGHT;*/
 
     public static Item MEDIUM_STONE_LANTERN_ITEM;
     public static Item MEDIUM_STONE_LANTERN_TOP_DARK_ITEM;
@@ -91,8 +91,8 @@ public class BlockFile {
     public static Item MINI_STONE_LANTERN_LIGHT_ITEM;
     public static Item SMALLFOOT_STONE_LANTERN_DARK_ITEM;
     public static Item SMALLFOOT_STONE_LANTERN_LIGHT_ITEM;
-    /*public static Item SMALLPROOF_STONE_LANTERN_DARK_ITEM;
-    public static Item SMALLPROOF_STONE_LANTERN_LIGHT_ITEM;*/
+    public static Item SMALLPROOF_STONE_LANTERN_DARK_ITEM;
+    public static Item SMALLPROOF_STONE_LANTERN_LIGHT_ITEM;
     public static Item SMALLNOPROOF_STONE_LANTERN_DARK_ITEM;
     public static Item SMALLNOPROOF_STONE_LANTERN_LIGHT_ITEM;
     public static Item MININOPROOF_STONE_LANTERN_DARK_ITEM;
@@ -132,8 +132,8 @@ public class BlockFile {
     public static Item BROAD_STONE_LANTERN_BLOCK_LIGHT_ITEM;
     public static Item MID_STONE_LANTERN_BLOCK_DARK_ITEM;
     public static Item MID_STONE_LANTERN_BLOCK_LIGHT_ITEM;
-    public static Item THIN_STONE_LANTERN_BLOCK_DARK_ITEM;
-    public static Item THIN_STONE_LANTERN_BLOCK_LIGHT_ITEM;
+    /*public static Item THIN_STONE_LANTERN_BLOCK_DARK_ITEM;
+    public static Item THIN_STONE_LANTERN_BLOCK_LIGHT_ITEM;*/
 
     public static void register() {
 
@@ -214,13 +214,13 @@ public class BlockFile {
         SMALLFOOT_STONE_LANTERN_LIGHT = registerBlock("smallfoot_stone_lantern_light", AbstractBlock.Settings.create().mapColor(MapColor.STONE_GRAY).sounds(BlockSoundGroup.STONE).strength(1.0F, 10.0F).luminance((state) -> 14).nonOpaque().pistonBehavior(PistonBehavior.DESTROY).solidBlock((state, world, pos) -> false), SmallfootStoneLanternLightBlock::new);
         SMALLFOOT_STONE_LANTERN_LIGHT_ITEM = registerBlockItem("smallfoot_stone_lantern_light", SMALLFOOT_STONE_LANTERN_LIGHT);
 
-        /*
+
         SMALLPROOF_STONE_LANTERN_DARK = registerBlock("smallproof_stone_lantern_dark", AbstractBlock.Settings.create().mapColor(MapColor.STONE_GRAY).sounds(BlockSoundGroup.STONE).strength(1.0F, 10.0F).nonOpaque().pistonBehavior(PistonBehavior.DESTROY).solidBlock((state, world, pos) -> false), SmallproofStoneLanternDarkBlock::new);
         SMALLPROOF_STONE_LANTERN_DARK_ITEM = registerBlockItem("smallproof_stone_lantern_dark", SMALLPROOF_STONE_LANTERN_DARK);
 
         SMALLPROOF_STONE_LANTERN_LIGHT = registerBlock("smallproof_stone_lantern_light", AbstractBlock.Settings.create().mapColor(MapColor.STONE_GRAY).sounds(BlockSoundGroup.STONE).strength(1.0F, 10.0F).luminance((state) -> 14).nonOpaque().pistonBehavior(PistonBehavior.DESTROY).solidBlock((state, world, pos) -> false), SmallproofStoneLanternLightBlock::new);
         SMALLPROOF_STONE_LANTERN_LIGHT_ITEM = registerBlockItem("smallproof_stone_lantern_light", SMALLPROOF_STONE_LANTERN_LIGHT);
-*/
+
 
         SMALLNOPROOF_STONE_LANTERN_DARK = registerBlock("smallnoproof_stone_lantern_dark", AbstractBlock.Settings.create().mapColor(MapColor.STONE_GRAY).sounds(BlockSoundGroup.STONE).strength(1.0F, 10.0F).nonOpaque().pistonBehavior(PistonBehavior.DESTROY).solidBlock((state, world, pos) -> false), SmallnoproofStoneLanternDarkBlock::new);
         SMALLNOPROOF_STONE_LANTERN_DARK_ITEM = registerBlockItem("smallnoproof_stone_lantern_dark", SMALLNOPROOF_STONE_LANTERN_DARK);
@@ -317,13 +317,13 @@ public class BlockFile {
         MID_STONE_LANTERN_BLOCK_LIGHT_ITEM = registerBlockItem("mid_stone_lantern_block_light", MID_STONE_LANTERN_BLOCK_LIGHT);
 
 
-        THIN_STONE_LANTERN_BLOCK_DARK = registerBlock("thin_stone_lantern_block_dark", AbstractBlock.Settings.create().mapColor(MapColor.STONE_GRAY).sounds(BlockSoundGroup.STONE).strength(1.0F, 10.0F).nonOpaque().pistonBehavior(PistonBehavior.DESTROY).solidBlock((state, world, pos) -> false), ThinStoneLanternBlockDarkBlock::new);
+        /*THIN_STONE_LANTERN_BLOCK_DARK = registerBlock("thin_stone_lantern_block_dark", AbstractBlock.Settings.create().mapColor(MapColor.STONE_GRAY).sounds(BlockSoundGroup.STONE).strength(1.0F, 10.0F).nonOpaque().pistonBehavior(PistonBehavior.DESTROY).solidBlock((state, world, pos) -> false), ThinStoneLanternBlockDarkBlock::new);
         THIN_STONE_LANTERN_BLOCK_DARK_ITEM = registerBlockItem("thin_stone_lantern_block_dark", THIN_STONE_LANTERN_BLOCK_DARK);
 
 
         THIN_STONE_LANTERN_BLOCK_LIGHT = registerBlock("thin_stone_lantern_block_light", AbstractBlock.Settings.create().mapColor(MapColor.STONE_GRAY).sounds(BlockSoundGroup.STONE).strength(1.0F, 10.0F).luminance((state) -> 15).nonOpaque().pistonBehavior(PistonBehavior.DESTROY).solidBlock((state, world, pos) -> false), ThinStoneLanternBlockLightBlock::new);
         THIN_STONE_LANTERN_BLOCK_LIGHT_ITEM = registerBlockItem("thin_stone_lantern_block_light", THIN_STONE_LANTERN_BLOCK_LIGHT);
-    }
+    */}
 
     private static Block registerBlock(String name, AbstractBlock.Settings baseSettings, java.util.function.Function<AbstractBlock.Settings, Block> blockFactory) {
         RegistryKey<Block> blockKey = RegistryKey.of(RegistryKeys.BLOCK, Identifier.of(ReddensstonelanternMod.MODID, name));